fix(acp): move workspace switch button to AcpChatViewHeader
The button was incorrectly added to DefaultChatViewHeader which is not
used in ACP mode. Moved to AcpChatViewHeader where the ACP-specific
header is rendered.
